Le massive export update
