+++ .gitignore
@@ -4,5 +4,3 @@ node_modules
 Build
 foundry
 styles/daggerheart.css
-daggerheart.js
-daggerheart.js.map
