+++ module/_module.mjs
@@ -1,0 +1,9 @@
+export * as applications from './applications/_module.mjs';
+export * as canvas from './canvas/_module.mjs';
+export * as config from './config/_module.mjs';
+export * as data from './data/_module.mjs';
+export * as dice from './dice/_module.mjs';
+export * as documents from './documents/_module.mjs';
+export * as enrichers from './enrichers/_module.mjs';
+export * as helpers from './helpers/_module.mjs';
+export * as systemRegistration from './systemRegistration/_module.mjs';
