+++ module/applications/characterCreation/_module.mjs
@@ -1,0 +1,1 @@
+export { default as CharacterCreation } from './characterCreation.mjs';
