+++ module/applications/dialogs/_module.mjs
@@ -1,0 +1,9 @@
+export { default as BeastformDialog } from './beastformDialog.mjs';
+export { default as costSelectionDialog } from './costSelectionDialog.mjs';
+export { default as d20RollDialog } from './d20RollDialog.mjs';
+export { default as DamageDialog } from './damageDialog.mjs';
+export { default as DamageReductionDialog } from './damageReductionDialog.mjs';
+export { default as DamageSelectionDialog } from './damageSelectionDialog.mjs';
+export { default as DeathMove } from './deathMove.mjs';
+export { default as Downtime } from './downtime.mjs';
+export { default as OwnershipSelection } from './ownershipSelection.mjs';
