+++ module/applications/sheets-configs/activeEffectConfig.mjs
