+++ module/applications/sheets/actors/_module.mjs
@@ -1,0 +1,4 @@
+export { default as Adversary } from './adversary.mjs';
+export { default as Character } from './character.mjs';
+export { default as Companion } from './companion.mjs';
+export { default as Environment } from './environment.mjs';
