+++ module/applications/sheets/actors/character.mjs
@@ -6,7 +6,7 @@ import { abilities } from '../../../config/actorConfig.mjs';
 import DhCharacterlevelUp from '../../levelup/characterLevelup.mjs';
 import DhCharacterCreation from '../../characterCreation/characterCreation.mjs';
 import FilterMenu from '../../ux/filter-menu.mjs';
-import DHActionConfig from '../../../config/Action.mjs';
+import DHActionConfig from '../../sheets-configs/action-config.mjs';
 
 const { ActorSheetV2 } = foundry.applications.sheets;
 const { TextEditor } = foundry.applications.ux;
