+++ module/applications/sheets/api/base-item.mjs
@@ -1,6 +1,6 @@
 import DHApplicationMixin from './application-mixin.mjs';
-import { actionsTypes } from '../../../data/_module.mjs';
-import DHActionConfig from '../../../config/Action.mjs';
+import DHActionConfig from '../../sheets-configs/action-config.mjs';
+import { actionsTypes } from '../../../data/action/_module.mjs';
 
 const { ItemSheetV2 } = foundry.applications.sheets;
