+++ module/applications/sheets/items/_module.mjs
@@ -1,0 +1,11 @@
+export { default as Ancestry } from './ancestry.mjs';
+export { default as Armor } from './armor.mjs';
+export { default as Beastform } from './beastform.mjs';
+export { default as Class } from './class.mjs';
+export { default as Community } from './community.mjs';
+export { default as Consumable } from './consumable.mjs';
+export { default as DomainCard } from './domainCard.mjs';
+export { default as Feature } from './feature.mjs';
+export { default as Miscellaneous } from './miscellaneous.mjs';
+export { default as Subclass } from './subclass.mjs';
+export { default as Weapon } from './weapon.mjs';
