+++ module/applications/ui/_module.mjs
@@ -1,6 +1,4 @@
-import { default as DhChatLog } from './chatLog.mjs';
-import { default as DhCombatTracker } from './combatTracker.mjs';
-import * as DhCountdowns from './countdowns.mjs';
-import { default as DhFearTracker } from './fearTracker.mjs';
-
-export { DhChatLog, DhCombatTracker, DhCountdowns, DhFearTracker };
+export { default as DhChatLog } from './chatLog.mjs';
+export { default as DhCombatTracker } from './combatTracker.mjs';
+export * as DhCountdowns from './countdowns.mjs';
+export { default as DhFearTracker } from './fearTracker.mjs';
