+++ module/applications/ux/_module.mjs
@@ -1,1 +1,2 @@
 export { default as FilterMenu } from './filter-menu.mjs';
+export { default as ContextMenu } from './contextMenu.mjs';
