+++ module/config/_module.mjs
@@ -1,25 +1,10 @@
-import * as action from './Action.mjs';
-import * as actionConfig from './actionConfig.mjs';
-import * as actorConfig from './actorConfig.mjs';
-import * as domainConfig from './domainConfig.mjs';
-import * as effectConfig from './effectConfig.mjs';
-import * as flagsConfig from './flagsConfig.mjs';
-import * as generalConfig from './generalConfig.mjs';
-import * as hooksConfig from './hooksConfig.mjs';
-import * as itemConfig from './itemConfig.mjs';
-import * as settingsConfig from './settingsConfig.mjs';
-import * as systemConfig from './system.mjs';
-
-export {
-    action,
-    actionConfig,
-    actorConfig,
-    domainConfig,
-    effectConfig,
-    flagsConfig,
-    generalConfig,
-    hooksConfig,
-    itemConfig,
-    settingsConfig,
-    systemConfig
-};
+export * as actionConfig from './actionConfig.mjs';
+export * as actorConfig from './actorConfig.mjs';
+export * as domainConfig from './domainConfig.mjs';
+export * as effectConfig from './effectConfig.mjs';
+export * as flagsConfig from './flagsConfig.mjs';
+export * as generalConfig from './generalConfig.mjs';
+export * as hooksConfig from './hooksConfig.mjs';
+export * as itemConfig from './itemConfig.mjs';
+export * as settingsConfig from './settingsConfig.mjs';
+export * as systemConfig from './system.mjs';
