+++ module/data/_module.mjs
@@ -1,9 +1,9 @@
 export { default as DhCombat } from './combat.mjs';
 export { default as DhCombatant } from './combatant.mjs';
 
-export * as actors from './actor/_module.mjs';
-export * as items from './item/_module.mjs';
-export { actionsTypes } from './action/_module.mjs';
-export * as messages from './chat-message/_modules.mjs';
-export * as fields from './fields/_module.mjs';
+export * as actions from './action/_module.mjs';
 export * as activeEffects from './activeEffect/_module.mjs';
+export * as actors from './actor/_module.mjs';
+export * as chatMessages from './chat-message/_modules.mjs';
+export * as fields from './fields/_module.mjs';
+export * as items from './item/_module.mjs';
