+++ module/data/action/_module.mjs
@@ -1,25 +1,20 @@
-import {
-    DHAttackAction,
-    DHBaseAction,
-    DhBeastformAction,
-    DHDamageAction,
-    DHEffectAction,
-    DHHealingAction,
-    DHMacroAction,
-    // DHResourceAction,
-    // DHSpellCastAction,
-    DHSummonAction
-} from './action.mjs';
+export * as ActionDice from './actionDice.mjs';
+import AttackAction from './attackAction.mjs';
+import BaseAction from './baseAction.mjs';
+import BeastformAction from './beastformAction.mjs';
+import DamageAction from './damageAction.mjs';
+import EffectAction from './effectAction.mjs';
+import HealingAction from './healingAction.mjs';
+import MacroAction from './macroAction.mjs';
+import SummonAction from './summonAction.mjs';
 
 export const actionsTypes = {
-    base: DHBaseAction,
-    attack: DHAttackAction,
-    // spellcast: DHSpellCastAction,
-    // resource: DHResourceAction,
-    damage: DHDamageAction,
-    healing: DHHealingAction,
-    summon: DHSummonAction,
-    effect: DHEffectAction,
-    macro: DHMacroAction,
-    beastform: DhBeastformAction
+    base: BaseAction,
+    attack: AttackAction,
+    damage: DamageAction,
+    healing: HealingAction,
+    summon: SummonAction,
+    effect: EffectAction,
+    macro: MacroAction,
+    beastform: BeastformAction
 };
