+++ module/data/actor/environment.mjs
@@ -1,4 +1,3 @@
-import { environmentTypes } from '../../config/actorConfig.mjs';
 import BaseDataActor from './base.mjs';
 import ActionField from '../fields/actionField.mjs';
 import ForeignDocumentUUIDArrayField from '../fields/foreignDocumentUUIDArrayField.mjs';
@@ -21,7 +20,7 @@ export default class DhEnvironment extends BaseDataActor {
                 choices: CONFIG.DH.GENERAL.tiers,
                 initial: CONFIG.DH.GENERAL.tiers.tier1.id
             }),
-            type: new fields.StringField({ choices: environmentTypes }),
+            type: new fields.StringField({ choices: CONFIG.DH.ACTOR.environmentTypes }),
             description: new fields.StringField(),
             impulses: new fields.StringField(),
             difficulty: new fields.NumberField({ required: true, initial: 11, integer: true }),
