+++ module/data/fields/actionField.mjs
@@ -1,9 +1,9 @@
-// import { actionsTypes } from '../action/_module.mjs';
+import { actionsTypes } from '../action/_module.mjs';
 
 // Temporary Solution
 export default class ActionField extends foundry.data.fields.ObjectField {
     getModel(value) {
-        return game.system.api.models.actionsTypes[value.type] ?? game.system.api.models.actionsTypes.attack;
+        return actionsTypes[value.type] ?? actionsTypes.attack;
     }
 
     /* -------------------------------------------- */
