+++ module/dice/_module.mjs
@@ -1,0 +1,5 @@
+export { default as D20Roll } from './d20Roll.mjs';
+export { default as DamageRoll } from './damageRoll.mjs';
+export { default as DHRoll } from './dhRoll.mjs';
+export { default as DualityDie } from './dualityDie.mjs';
+export { default as DualityRoll } from './dualityRoll.mjs';
