+++ module/documents/_module.mjs
@@ -3,5 +3,4 @@ export { default as DHItem } from './item.mjs';
 export { default as DhpCombat } from './combat.mjs';
 export { default as DhActiveEffect } from './activeEffect.mjs';
 export { default as DhChatMessage } from './chatMessage.mjs';
-export * as DhRoll from '../dice/dhRolls.mjs';
 export { default as DhTooltipManager } from './tooltipManager.mjs';
