+++ module/helpers/_module.mjs
@@ -1,0 +1,2 @@
+export { default as HandlebarsHelper } from './handlebarsHelper.mjs';
+export * as utils from './utils.mjs';
