+++ module/systemRegistration/_module.mjs
@@ -1,5 +1,3 @@
-import { preloadHandlebarsTemplates as handlebarsRegistration } from './handlebars.mjs';
-import * as settingsRegistration from './settings.mjs';
-import * as socketRegistration from './socket.mjs';
-
-export { handlebarsRegistration, settingsRegistration, socketRegistration };
+export { preloadHandlebarsTemplates as handlebarsRegistration } from './handlebars.mjs';
+export * as settingsRegistration from './settings.mjs';
+export * as socketRegistration from './socket.mjs';
